+++ src/index.ts
@@ -39,7 +39,6 @@ createCommand(process.argv, globalThis, logger)
         break
       }
     } else if (err instanceof CommanderError) {
-      // Catched by commander library
       return exit(1)
     } else {
       logger.error(`Unhandled error: ${err instanceof TypeError ? err.message : err}`)