+++ tests/integration/commands/publish.test.ts
@@ -180,48 +180,6 @@ describe('E2E: publish project', () => {
     })
   })
 
-  describe('only files without package.json', () => {
-    it('needs --scope', async () => {
-      const resources = ['a.js']
-      const { repositoryCtx } = await createPackageCtx({ resources })
-      const filesPath = resources.map(res => absoluteResolve(repositoryCtx.name, res))
-
-      await expect(createCommand(
-        buildCommandArguments([...baseCommand, ...baseArgs, ...filesPath]),
-        global,
-        loggerStub
-      )).to.be.eventually.rejectedWith('No scope was matched in package.json `name` field or in --scope')
-
-      await repositoryCtx.cleanup()
-    })
-
-    it('should push files', async () => {
-      const scope = '__test/foo/dir/files'
-      const scopeDir: RelPath = `./${scope}`
-      const resources = ['a.js', 'b.txt', 'foo', 'bar']
-      const { repositoryCtx } = await createPackageCtx({ resources })
-      const filesPath = resources.map(res => absoluteResolve(repositoryCtx.name, res))
-
-      await expect(createCommand(
-        buildCommandArguments([...baseCommand, ...baseArgs, '--scope', scope, ...filesPath]),
-        global,
-        loggerStub
-      )).to.be.eventually.fulfilled
-
-
-      try {
-        await client.list(scopeDir)
-      } catch (error) {
-        console.log(error)
-      }
-      await expect(client.list(scopeDir))
-        .to.eventually.be.fulfilled.and.to.have.length(resources.length)
-
-
-      await repositoryCtx.cleanup()
-    })
-  })
-
   describe('with --override-version', () => {
     it('should push empty senver package', async () => {
       const { repositoryCtx, createCdnPath } = await createPackageCtx()