+++ CHANGELOG.md
@@ -6,3 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+### Added
+
+- command `update` added to push files to a remote CDN location without the npm semver PUT logic
+
+### BREAKING CHANGE
+
+- `publish` does not support anymore the [...files] argument. The behavior is moved to the `update` command